rename: stories 폴더 및 파일 이름을 epilogue로 변경 (#9)
